refactor(Onboarding): 훅 및 컴포넌트 분리
